feat: context on home page
