+++ src/pages/Home/components/CountDown/index.tsx
@@ -1,0 +1,77 @@
+import React, { useContext, useEffect } from 'react'
+import { differenceInSeconds } from 'date-fns'
+
+import { setDocumentTitle } from '../../../../utils/documentTitle'
+
+import { CountDownContainer, Separetor } from './styles'
+import { CyclesContext } from '../..'
+
+const CountDown: React.FC = () => {
+  const {
+    activeCycle,
+    activeCycleId,
+    markCurrentCycleAsFinished,
+    amountSecondsPasses,
+    setSecondsPassed,
+  } = useContext(CyclesContext)
+
+  const totalSeconds = activeCycle ? activeCycle.minutesAmount * 60 : 0
+  const currentSeconds = activeCycle ? totalSeconds - amountSecondsPasses : 0
+
+  const minutesAmount = Math.floor(currentSeconds / 60)
+  const secondsAmount = currentSeconds % 60
+
+  const minutes = String(minutesAmount).padStart(2, '0')
+  const seconds = String(secondsAmount).padStart(2, '0')
+
+  useEffect(() => {
+    if (activeCycle) {
+      setDocumentTitle(`${activeCycle.task} - ${minutes}:${seconds}`)
+    }
+  }, [activeCycle, minutes, seconds])
+
+  useEffect(() => {
+    let interval: number
+
+    if (activeCycle) {
+      interval = setInterval(() => {
+        const secondsDifference = differenceInSeconds(
+          new Date(),
+          activeCycle.startDate,
+        )
+
+        if (secondsDifference >= totalSeconds) {
+          markCurrentCycleAsFinished()
+
+          setDocumentTitle('', true)
+          setSecondsPassed(totalSeconds)
+          clearInterval(interval)
+        } else {
+          setSecondsPassed(secondsDifference)
+        }
+      }, 1000)
+    }
+
+    return () => {
+      clearInterval(interval)
+    }
+  }, [
+    activeCycle,
+    activeCycleId,
+    markCurrentCycleAsFinished,
+    setSecondsPassed,
+    totalSeconds,
+  ])
+
+  return (
+    <CountDownContainer>
+      <span>{minutes[0]}</span>
+      <span>{minutes[1]}</span>
+      <Separetor>:</Separetor>
+      <span>{seconds[0]}</span>
+      <span>{seconds[1]}</span>
+    </CountDownContainer>
+  )
+}
+
+export default CountDown
