+++ src/pages/Home/components/index.ts
@@ -1,0 +1,2 @@
+export { default as CountDown } from './CountDown'
+export { default as NewCycleForm } from './NewCycleForm'
